Fix mouse wheel down on windows.

+++ src/arch/InputHandler/InputHandler_DirectInput.cpp
@@ -667,7 +667,9 @@ void InputHandler_DInput::UpdateBuffered( DIDevice &device, const RageTimer &tm 
 					DeviceButton up = DeviceButton_Invalid, down = DeviceButton_Invalid;
 					if(dev == DEVICE_MOUSE)
 					{
-						float l = static_cast<float>(evtbuf[i].dwData);
+						// dwData is unsigned, so it has to be converted to int first to
+						// make it signed.
+						float l = float(int32_t(evtbuf[i].dwData));
 						POINT cursorPos;
 						GetCursorPos(&cursorPos);
 						// convert screen coordinates to client
